requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.35.0
mcp-use>=0.1.0
fastmcp>=0.1.0
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
langchain-openai>=0.1.0
